requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
decorator==5.0.9
Flask==2.0.1
Flask-RESTful==0.3.9
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.1.5
patsy==0.5.1
py==1.10.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
retry==0.9.2
scipy==1.7.1
six==1.16.0
statsmodels==0.12.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
